Added Noodle Thickness at preferences
The default value for the thickness of the noodle can be modified from xNode preferences.

This value is only effective if the user does not make an override from the EditorGraph at GetNoodleThickness(...).

Many users do not like to edit from the script, this will help in the creation of my new layout template, similar to the new graphView from unity